处理缩略图吃空间，页面等待脱敏完成改为定时任务处理之后

+++ bigdata-es/src/main/java/qed/bigdata/es/service/TagService.java
@@ -32,9 +32,9 @@ public interface TagService {
     /**
      * 做实际的脱敏操作
      * @param tag
-     * @return 返回被脱敏序列的数量
+     * @return 返回被脱敏任务提交状态，
      */
-    Long doDesensitize(String tag);
+    Integer doDesensitize(String tag);
 
     /**
      * 该方法查询所有已经存在的所有标签并返回每个标签的的序列数量。
